+++ spec/spec_helper.rb
@@ -4,6 +4,19 @@ require "bundler/setup"
 
 Bundler.require(:default, :development)
 
+if ENV["CI"] || ENV["COVERAGE"]
+  require "simplecov"
+  require "simplecov-console"
+
+  SimpleCov.formatter = SimpleCov::Formatter::Console
+  SimpleCov.start do
+    add_filter("lib/pusher-fake/support")
+    add_filter("spec/support")
+    enable_coverage :branch
+    minimum_coverage line: 100, branch: 100
+  end
+end
+
 Dir[File.expand_path("support/**/*.rb", __dir__)].sort.each do |file|
   require file
 end