+++ spec/support/coveralls.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-if ENV["CI"]
-  require "coveralls"
-
-  Coveralls.wear_merged! do
-    add_filter "spec/"
-  end
-end